+++ src/libs/ValueSchemaError.ts
@@ -27,7 +27,7 @@ export class ValueSchemaError extends Error
 	 * @param err error to check
 	 * @returns Yes/No
 	 */
-	static instances(err: unknown): err is ValueSchemaError
+	static is(err: unknown): err is ValueSchemaError
 	{
 		return err instanceof ValueSchemaError;
 	}