+++ src/schemas/object.ts
@@ -1,3 +1,4 @@
+import * as converter from "../appliers/converter";
 import * as ifUndefined from "../appliers/ifUndefined";
 import * as ifEmptyString from "../appliers/ifEmptyString";
 import * as ifNull from "../appliers/ifNull";
@@ -7,6 +8,7 @@ import * as type from "../appliers/object/type";
 import {BaseSchema} from "../libs/BaseSchema";
 
 type OptionsForObject =
+	converter.Options<object> |
 	ifUndefined.Options<object> |
 	ifEmptyString.Options<object> |
 	ifNull.Options<object> |
@@ -23,6 +25,7 @@ class ObjectSchema extends BaseSchema<object>
 			ifEmptyString.applyTo,
 			schema.applyTo,
 			type.applyTo,
+			converter.applyTo,
 		]);
 	}
 }