+++ test/schemas/string.test.ts
@@ -9,8 +9,8 @@ import vs from "value-schema";
 	describe("only", testOnly);
 	describe("minLength", testMinLength);
 	describe("maxLength", testMaxLength);
-	describe("caseConverter", testCaseConverter);
 	describe("pattern", testPattern);
+	describe("converter", testConverter);
 }
 
 /**
@@ -282,36 +282,6 @@ function testMaxLength(): void
 				maxLength: 8,
 			}).applyTo("123456789");
 		}).toThrow(vs.CAUSE.MAX_LENGTH);
-	});
-}
-
-/**
- * case converter
- */
-function testCaseConverter(): void
-{
-	describe("should be adjusted", () =>
-	{
-		expect(
-			vs.string({
-				caseConverter: vs.STRING.CASE_CONVERTER.LOWER,
-			}).applyTo("123ABCxyz")
-		).toEqual("123abcxyz");
-
-		expect(
-			vs.string({
-				caseConverter: vs.STRING.CASE_CONVERTER.UPPER,
-			}).applyTo("123ABCxyz")
-		).toEqual("123ABCXYZ");
-	});
-	it("should cause error(s)", () =>
-	{
-		expect(() =>
-		{
-			vs.string({
-				caseConverter: "" as any, // eslint-disable-line @typescript-eslint/no-explicit-any
-			}).applyTo("123ABCxyz");
-		}).toThrow(vs.CAUSE.CASE_CONVERTER);
 	});
 }
 
@@ -702,5 +672,52 @@ function testPatternOthers(): void
 				pattern: vs.STRING.PATTERN.EMAIL,
 			}).applyTo("john..doe@example.com");
 		}).toThrow(vs.CAUSE.PATTERN);
+	});
+}
+
+/**
+ * converter
+ */
+function testConverter(): void
+{
+	it("should be adjusted", () =>
+	{
+		expect(
+			vs.string({
+				converter: (value) =>
+				{
+					return value.toLowerCase();
+				},
+			}).applyTo("123ABCxyz")
+		).toEqual("123abcxyz");
+
+		expect(
+			vs.string({
+				converter: (value, fail) =>
+				{
+					if(value.length >= 9)
+					{
+						fail();
+					}
+					return value.padStart(8, " ");
+				},
+			}).applyTo("test")
+		).toEqual("    test");
+	});
+	it("should cause error(s)", () =>
+	{
+		expect(() =>
+		{
+			vs.string({
+				converter: (value, fail) =>
+				{
+					if(value.length >= 9)
+					{
+						fail();
+					}
+					return value.padStart(8, " ");
+				},
+			}).applyTo("123ABCxyz");
+		}).toThrow(vs.CAUSE.CONVERTER);
 	});
 }