+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+* `isValueSchemaError(err)` to `ValueSchemaError.underlies(err)`
+
 ## [3.0.0-rc.2] - 2020-02-16
 
 ### Fixed